feat: provide approveAllCollectionItems as standalone function
Without this patch there's only the option to send both approval
transactions in a bundle. This can be a problem when the Safe has one of
the approvals but not both. Surely this is an edge case, but it did
happen during development. The problem, here, is that the approval of
the transfer manager will reject if it's already done. This could lead
to users being stuck.